makefile and pg targets (#245)
