+++ cmd/pg/backup_fetch.go
@@ -1,4 +1,4 @@
-package cmd
+package pg
 
 import (
 	"github.com/spf13/cobra"
@@ -23,5 +23,5 @@ var backupFetchCmd = &cobra.Command{
 }
 
 func init() {
-	RootCmd.AddCommand(backupFetchCmd)
+	PgCmd.AddCommand(backupFetchCmd)
 }
