+++ cmd/pg/wal_prefetch.go
@@ -1,4 +1,4 @@
-package cmd
+package pg
 
 import (
 	"github.com/spf13/cobra"
@@ -25,5 +25,5 @@ var walPrefetchCmd = &cobra.Command{
 }
 
 func init() {
-	RootCmd.AddCommand(walPrefetchCmd)
+	PgCmd.AddCommand(walPrefetchCmd)
 }
