+++ cmd/pg/wal_push.go
@@ -1,4 +1,4 @@
-package cmd
+package pg
 
 import (
 	"github.com/spf13/cobra"
@@ -23,5 +23,5 @@ var walPushCmd = &cobra.Command{
 }
 
 func init() {
-	RootCmd.AddCommand(walPushCmd)
+	PgCmd.AddCommand(walPushCmd)
 }
