+++ main/configure.go
@@ -1,0 +1,52 @@
+package config
+
+import (
+	"fmt"
+	"github.com/mitchellh/go-homedir"
+	"github.com/spf13/viper"
+	"github.com/wal-g/wal-g/internal"
+	"github.com/wal-g/wal-g/internal/tracelog"
+	"os"
+)
+
+var cfgFile string
+
+func Configure() {
+	err := internal.ConfigureLogging()
+	if err != nil {
+		tracelog.ErrorLogger.Println("Failed to configure logging.")
+		tracelog.ErrorLogger.FatalError(err)
+	}
+
+	err = internal.ConfigureLimiters()
+	if err != nil {
+		tracelog.ErrorLogger.Println("Failed to configure limiters")
+		tracelog.ErrorLogger.FatalError(err)
+	}
+}
+
+// initConfig reads in config file and ENV variables if set.
+func InitConfig() {
+	if cfgFile != "" {
+		// Use config file from the flag.
+		viper.SetConfigFile(cfgFile)
+	} else {
+		// Find home directory.
+		home, err := homedir.Dir()
+		if err != nil {
+			fmt.Println(err)
+			os.Exit(1)
+		}
+
+		// Search config in home directory with name ".wal-g" (without extension).
+		viper.AddConfigPath(home)
+		viper.SetConfigName(".wal-g")
+	}
+
+	viper.AutomaticEnv() // read in environment variables that match
+
+	// If a config file is found, read it in.
+	if err := viper.ReadInConfig(); err == nil {
+		fmt.Println("Using config file:", viper.ConfigFileUsed())
+	}
+}
